# backend/openapi_spec.yaml
openapi: 3.0.3
info:
  title: Geo Locator API - ЛЦТ 2025
  description: |
    Система геолокации и анализа нарушений с интеграцией датасета заказчика.
    
    **Основные возможности:**
    - Двойная ИИ система детекции (YOLO + Mistral AI)
    - Интеграция с эталонной базой датасета (11,239 изображений)
    - Российские геосервисы (Яндекс, 2GIS, Роскосмос, OSM)
    - Производительность: 1000 фото за 3 часа
    - Экспорт в XLSX согласно требованиям ТЗ
    
  version: 1.0.0
  contact:
    name: Geo Locator Team
    email: support@geolocator.ru
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001
    description: Development server
  - url: https://api.geolocator.ru
    description: Production server
...
paths:
...
  /api/auth/login:
    post:
      summary: Авторизация пользователя
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "admin"
                password:
                  type: string
                  example: "password"
              required: [username, password]
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT